+++ llvm/include/llvm/Support/RISCVISAUtils.h
@@ -14,7 +14,6 @@
 #define LLVM_SUPPORT_RISCVISAUTILS_H
 
 #include "llvm/ADT/StringRef.h"
-#include "llvm/ADT/StringSwitch.h"
 #include "llvm/Support/Compiler.h"
 #include <map>
 #include <string>
@@ -43,21 +42,6 @@ struct ExtensionComparator {
 /// in canonical order of extension.
 using OrderedExtensionMap =
     std::map<std::string, ExtensionVersion, ExtensionComparator>;
-
-enum class ZicfilpLabelSchemeKind {
-  Invalid,
-  Unlabeled,
-  FuncSig,
-};
-
-// See clang::getCFBranchLabelSchemeFlagVal() for possible CFBranchLabelScheme.
-inline ZicfilpLabelSchemeKind
-getZicfilpLabelScheme(const StringRef CFBranchLabelScheme) {
-  return StringSwitch<ZicfilpLabelSchemeKind>(CFBranchLabelScheme)
-      .Case("unlabeled", ZicfilpLabelSchemeKind::Unlabeled)
-      .Case("func-sig", ZicfilpLabelSchemeKind::FuncSig)
-      .Default(ZicfilpLabelSchemeKind::Invalid);
-}
 
 } // namespace RISCVISAUtils
